+++ packages/deploy-helpers/src/deploy/helpers/create-worker-upload-form.ts
@@ -80,6 +80,7 @@ export function createWorkerUploadForm(
 		assets,
 		observability,
 		cache,
+		package_dependencies,
 	} = worker;
 
 	const assetConfig: AssetConfigMetadata = {
@@ -849,6 +850,7 @@ export function createWorkerUploadForm(
 		}),
 		...(observability && { observability }),
 		...(cache && { cache_options: cache }),
+		...(package_dependencies?.length && { package_dependencies }),
 	};
 
 	if (options?.unsafe?.metadata !== undefined) {