+++ packages/miniflare/src/workers/workerd-unsafe.d.ts
@@ -1,0 +1,14 @@
+declare module "workerd:unsafe" {
+	export interface DurableObjectEvictionOptions {
+		webSockets?: "close" | "hibernate";
+	}
+
+	function evict(
+		stub: DurableObjectStub,
+		options?: DurableObjectEvictionOptions
+	): Promise<void>;
+
+	export default {
+		evict,
+	};
+}