+++ packages/wrangler/src/pages/run-workers-deploy.ts
@@ -52,7 +52,6 @@ export async function runPagesToWorkersDeploy(
 		routes: undefined,
 		domains: undefined,
 		metafile: undefined,
-		legacyEnv: undefined,
 		logpush: undefined,
 		oldAssetTtl: undefined,
 		dispatchNamespace: undefined,